requirements.txt
httpx
redis
python-dotenv